fix: layout 제거
